Move code into CheckIn.jsx
